# docs/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: OpenMSP / ApexMSP Control Plane API
  version: 1.0.0
  description: Multi-tenant MSP RMM and PSA Control Plane API specification.
servers:
  - url: /api/v1
    description: Control Plane v1 API
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PSATicket:
      type: object
      properties:
        id:
          type: string
        ticketNumber:
          type: string
        title:
          type: string
        description:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [new, in_progress, waiting_on_client, resolved, closed]
        category:
          type: string
        assignedTech:
          type: string
        createdAt:
          type: string
        slaDueDate:
          type: string
        slaBreached:
          type: boolean
        comments:
          type: array
          items:
            type: object
        timeEntries:
          type: array
          items:
            type: object
...
paths:
...
  /tickets:
    get:
      summary: List PSA tickets
      security:
        - bearerAuth: []
      parameters:
        - name: clientId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PSATicket'
    post:
      summary: Create ticket
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Ticket created
  /tickets/{id}:
    get:
      summary: Get ticket details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ticket details
    patch:
      summary: Update ticket status or fields
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ticket updated
  /tickets/{id}/comments:
    post:
      summary: Add comment to ticket
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Comment added
  /tickets/{id}/time:
    post:
      summary: Add time entry to ticket
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Time entry added
...